pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hermes"
version = "0.1.0"
description = "Local-first, memory-safe, LLM-powered document extraction engine."
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.9",
    "rich>=13.0",
    "pydantic>=2.0",
    "httpx>=0.27",
    "openpyxl>=3.1",
    "pymupdf>=1.24",
    "litellm>=1.30",
    "tenacity>=8.2",
]

[project.optional-dependencies]
ocr = [
    "surya-ocr>=0.4",
    "easyocr>=1.7",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
hermes = "hermes.cli:app_entry"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
